Switch to a more maintained "moment-element"
We really don't need much here, except for a bower.json, a git tag, and polymer 2.x compatibility. Our fork
of the bendavis78/moment-element provided that, but it doesn't get maintained in any way.

Instead use the element that has the name "moment-element" in the bower registry, and that does provide all of that.